+++ demos/serverless/src/handlers.js
@@ -17,6 +17,17 @@ const chime = new AWS.Chime({ region: 'us-east-1' });
 // Set the AWS SDK Chime endpoint. The global endpoint is https://service.chime.aws.amazon.com.
 chime.endpoint = new AWS.Endpoint(process.env.CHIME_ENDPOINT);
 
+const chimeRegional = new AWS.ChimeSDKMeetings({ region: 'us-east-1' });
+
+// return regional API just for Echo Reduction for now.
+function getClientForMeeting(meeting) {
+  if (meeting && meeting.Meeting && meeting.Meeting.MeetingFeatures && meeting.Meeting.MeetingFeatures.Audio &&
+    meeting.Meeting.MeetingFeatures.Audio.EchoReduction === 'AVAILABLE') {
+      return chimeRegional;
+  }
+  return chime;
+}
+
 // Read resource names from the environment
 const {
   MEETINGS_TABLE_NAME,
@@ -44,30 +55,37 @@ exports.join = async(event, context) => {
 
   // Look up the meeting by its title. If it does not exist, create the meeting.
   let meeting = await getMeeting(query.title);
+  
+  let client = getClientForMeeting(meeting);
+
   if (!meeting) {
-    const request = {
+    let request = {
       // Use a UUID for the client request token to ensure that any request retries
       // do not create multiple meetings.
       ClientRequestToken: uuidv4(),
-
+  
       // Specify the media region (where the meeting is hosted).
       // In this case, we use the region selected by the user.
       MediaRegion: query.region,
-
+  
       // Set up SQS notifications if being used
       NotificationsConfiguration: USE_EVENT_BRIDGE === 'false' ? { SqsQueueArn: SQS_QUEUE_ARN } : {},
-
+  
       // Any meeting ID you wish to associate with the meeting.
       // For simplicity here, we use the meeting title.
       ExternalMeetingId: query.title.substring(0, 64),
-
-      // Tags associated with the meeting. They can be used in cost allocation console
-      Tags: [
-        { Key: 'Department', Value: 'RND'}
-      ]
     };
+    if (query.ns_es === 'true') {
+      client = chimeRegional;
+      request.MeetingFeatures = {
+        Audio: {
+          // The EchoReduction parameter helps the user enable and use Amazon Echo Reduction.
+          EchoReduction: 'AVAILABLE'
+        }
+      };
+    } 
     console.info('Creating new meeting: ' + JSON.stringify(request));
-    meeting = await chime.createMeeting(request).promise();
+    meeting = await client.createMeeting(request).promise();
 
     // Store the meeting in the table using the meeting title as the key.
     await putMeeting(query.title, meeting);
@@ -75,7 +93,7 @@ exports.join = async(event, context) => {
 
   // Create new attendee for the meeting
   console.info('Adding new attendee');
-  const attendee = (await chime.createAttendee({
+  const attendee = (await client.createAttendee({
     // The meeting ID of the created meeting to add the attendee to
     MeetingId: meeting.Meeting.MeetingId,
 
@@ -99,15 +117,18 @@ exports.join = async(event, context) => {
 exports.end = async (event, context) => {
   // Fetch the meeting by title
   const meeting = await getMeeting(event.queryStringParameters.title);
+  let client = getClientForMeeting(meeting);
 
   // End the meeting. All attendee connections will hang up.
-  await chime.deleteMeeting({ MeetingId: meeting.Meeting.MeetingId }).promise();
+  await client.deleteMeeting({ MeetingId: meeting.Meeting.MeetingId }).promise();
   return response(200, 'application/json', JSON.stringify({}));
 };
 
 exports.start_transcription = async (event, context) => {
   // Fetch the meeting by title
   const meeting = await getMeeting(event.queryStringParameters.title);
+  let client = getClientForMeeting(meeting);
+
   const languageCode = event.queryStringParameters.language;
   const region = event.queryStringParameters.region;
   let transcriptionConfiguration = {};
@@ -163,7 +184,7 @@ exports.start_transcription = async (event, context) => {
   }
 
   // start transcription for the meeting
-  await chime.startMeetingTranscription({
+  await client.startMeetingTranscription({
     MeetingId: meeting.Meeting.MeetingId,
     TranscriptionConfiguration: transcriptionConfiguration
   }).promise();
@@ -173,9 +194,10 @@ exports.start_transcription = async (event, context) => {
 exports.stop_transcription = async (event, context) => {
   // Fetch the meeting by title
   const meeting = await getMeeting(event.queryStringParameters.title);
+  let client = getClientForMeeting(meeting);
 
   // stop transcription for the meeting
-  await chime.stopMeetingTranscription({
+  await client.stopMeetingTranscription({
     MeetingId: meeting.Meeting.MeetingId
   }).promise();
   return response(200, 'application/json', JSON.stringify({}));