+++ CHANGELOG.md
@@ -7,16 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add support for Echo Reduction when using Voice Focus.
 
 ### Removed
 
 ### Fixed
+
 - Switched `DefaultSimulcastUplinkPolicy` non-simulcast mode's enabled stream to mitigate Chromium not reconfiguring encoder on reconnect when using the top stream. This impacts only the non-simulcast mode of the simulcast policy when is when there are 2 or less participants (not 2 or less senders).
+
+### Changed
 
 ## [2.22.0] - 2021-11-18
 ### Added
 - Add documentation in video processor on how to add a customized professor with new image loading.
-
 - Adds support to live transcription for new features including PII content identification and redaction, partial results stabilization, and custom language models for Amazon Transcribe and PHI content identification for Amazon Transcribe Medical.
 
 ### Removed